refactor: remove CanvasBackups.tsx component and integrate BackupsDialog
- Deleted the CanvasBackups component and its associated files to streamline the codebase.
- Updated the MainMenu to use the new BackupsDialog component for managing canvas backups.
- Introduced new styles for the BackupsDialog to enhance the user interface and experience.
