+++ src/frontend/src/CustomEmbeddableRenderer.tsx
@@ -7,7 +7,6 @@ import {
   ControlButton,
   HtmlEditor,
   Editor,
-  CanvasBackups
 } from './pad';
 import { ActionButton } from './pad/buttons';
 
@@ -38,8 +37,6 @@ export const renderCustomEmbeddable = (
         />;
       case 'dashboard':
         return <Dashboard element={element} appState={appState} excalidrawAPI={excalidrawAPI} />;
-      case 'backups':
-        return <CanvasBackups element={element} appState={appState} excalidrawAPI={excalidrawAPI} />;
       default:
         return null;
     }
