--- src/frontend/src/pad/backups/index.ts
@@ -1,2 +1,0 @@
-export * from './CanvasBackups';
-export { default as CanvasBackups } from './CanvasBackups';
