+++ src/frontend/src/pad/index.ts
@@ -4,10 +4,8 @@ export * from './controls/StateIndicator';
 export * from './containers/Dashboard';
 export * from './buttons';
 export * from './editors';
-export * from './backups';
 
 // Default exports
 export { default as ControlButton } from './controls/ControlButton';
 export { default as StateIndicator } from './controls/StateIndicator';
 export { default as Dashboard } from './containers/Dashboard';
-export { default as CanvasBackups } from './backups/CanvasBackups';
