+++ src/frontend/src/ui/MainMenu.tsx
@@ -4,12 +4,11 @@ import type { ExcalidrawImperativeAPI } from '@atyrode/excalidraw/types';
 import type { MainMenu as MainMenuType } from '@atyrode/excalidraw';
 
 import { LogOut, SquarePlus, LayoutDashboard, SquareCode, Eye, Coffee, Grid2x2, User, Text, ArchiveRestore } from 'lucide-react';
-import BackupsModal from './BackupsModal';
+import BackupsModal from './BackupsDialog';
 import { capture } from '../utils/posthog';
 import { ExcalidrawElementFactory, PlacementMode } from '../lib/ExcalidrawElementFactory';
 import { useUserProfile } from "../api/hooks";
 import { queryClient } from "../api/queryClient";
-import Editor from '../pad/editors/Editor';
 
 interface MainMenuConfigProps {
   MainMenu: typeof MainMenuType;
@@ -95,19 +94,13 @@ export const MainMenuConfig: React.FC<MainMenuConfigProps> = ({
   };
 
   const [showBackupsModal, setShowBackupsModal] = useState(false);
-  const [isBackupsModalExiting, setIsBackupsModalExiting] = useState(false);
 
   const handleCanvasBackupsClick = () => {
     setShowBackupsModal(true);
-    setIsBackupsModalExiting(false);
-  };
-
-  const handleBackupsModalExitComplete = () => {
-    setShowBackupsModal(false);
   };
 
   const handleCloseBackupsModal = () => {
-    setIsBackupsModalExiting(true);
+    setShowBackupsModal(false);
   };
 
   const handleGridToggle = () => {
@@ -247,8 +240,6 @@ export const MainMenuConfig: React.FC<MainMenuConfigProps> = ({
       {showBackupsModal && (
         <BackupsModal
           excalidrawAPI={excalidrawAPI}
-          isExiting={isBackupsModalExiting}
-          onExitComplete={handleBackupsModalExitComplete}
           onClose={handleCloseBackupsModal}
         />
       )}
